refactor: AboutPage 컴포넌트 분리 및 JSDoc 문서화
